Removed Debezium Configuration.

+++ src/main/resources/db/migration/V1__init_schema.sql
@@ -1,1 +1,1 @@
-create table location (id bigserial not null , user_id uuid not null, latitude decimal, longitude decimal, created_on timestamp, primary key (id));
+create table location (id bigserial not null , user_id uuid not null, latitude double precision, longitude double precision, created_on timestamp, primary key (id));
